Move node utils package
